Change from throwing to returning string
Use object of type Error in suggestions.js instead of try-catch
(https://nodejs.org/en/knowledge/errors/what-is-try-catch/#but-wait-isn-t-it-node-js-convention-to-not-use-try-catch).

Instead of throwing an error, generate a string containing the error
message. When calling the validator (from suggestions.js) create and
handle a new error if said string is non-null.

Change `.forEach()` to `for-of` loops due to the former not terminating
early when using a return statement.